backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.2
python-dotenv==1.0.1
timezonefinder==6.5.0
pytz==2024.2
